started new class

+++ data structures class/BinarySearch.py
@@ -1,8 +1,10 @@
 pos = -1                                #this is setting a global variable
 
 def search(list, n):
+
     l = 0
     u = len(list)-1
+
     while l <= u:
         mid = (l + u) // 2
 
@@ -11,14 +13,14 @@ def search(list, n):
             return True
         else:
             if list[mid] < n:
-                l = mid+1
+                l = mid + 1
             else:
                 u = mid+1
     return False
 
 list = [2,5,16,27,46,83]                    #this is the list of elements
 
-n = float(input("enter a number: "))    #this ask the user to input a number to check if its in the list
+n = int(input("enter a number: "))      #this ask the user to input a number to check if its in the list
 
 if search(list, n):                     #this if statement says if our search finds an element tell us where it is located
     print("found at ", pos+1)
